# contracts/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: WSC Access Workbench API
  version: 2.2.0
  description: |
    wsc-contracts@2.2.0 — 接入端工作台 V1.4（座序图 L2 Top5、持久化用户、我的数据产品；SNAP-WSC-005）。
    统一响应包装：{ code, message, data, correlationId }；成功时 code="0"。
    RBAC：产品增/改/导入仅 PROVIDER；ADMIN 保留分类维护与用户管理；目录维护 ADMIN+PROVIDER（PROVIDER 仅 create_by=本人）。
    GET /catalog/products?mine=true 仅返回 create_by=当前用户的产品。
    GET /catalog/l2-distribution 返回二级目录产品计数（座序图）。
    ADMIN /admin/users CRUD 持久化 sys_user。
    产品编码格式（DEC-WSC-001）：`{DOMAIN}-{FEATURE}-{NNNN}`。
    认证：服务端会话（Cookie）；写 API 以服务端会话角色为准。
    （HOTFIX TASK-WSC-607：目录维护对 PROVIDER 开放本人产品；契约版本号未 bump，仍为 2.2.0。）
servers:
  - url: /api/v1
    description: V1 API base
...
security:
  - SessionCookie: []
paths:
...
  /chain/products/{productId}/versions:
    get:
      tags: [chain]
      summary: 产品上链版本列表（最新优先）
      operationId: listChainVersions
      parameters:
        - $ref: "#/components/parameters/ProductId"
      responses:
        "200":
          description: 版本列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseChainVersionList"
  /chain/versions/{versionId}/snapshot:
    get:
      tags: [chain]
      summary: 按 versionId 读取完整目录快照
      operationId: getChainSnapshot
      parameters:
        - name: versionId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: 不可变目录快照
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseChainSnapshot"
        "404":
          $ref: "#/components/responses/NotFound"
  /catalog/maintenance/entries:
    get:
      tags: [catalog]
      summary: 目录维护列表（条目≡数据产品；REQ-CAT-007 / TASK-WSC-607）
      description: |
        ADMIN 返回全量；PROVIDER 仅 create_by=当前会话用户；USER → 403 ERR_MAINTENANCE_FORBIDDEN。
      operationId: listMaintenanceEntries
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [ALL, MAINTAINED, PENDING]
            default: ALL
        - name: l1CategoryId
          in: query
          schema: { type: string }
        - name: l2CategoryId
          in: query
          schema: { type: string }
        - name: l3CategoryId
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, default: 1, minimum: 1 }
        - name: pageSize
          in: query
          schema: { type: integer, default: 20, minimum: 1, maximum: 100 }
      responses:
        "200":
          description: 分页列表（page/pageSize/total）
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseMaintenancePage"
        "403":
          $ref: "#/components/responses/Forbidden"
  /catalog/maintenance/entries/{productId}:
    put:
      tags: [catalog]
      summary: 单条目录维护关联（挂三级）
      description: |
        ADMIN 可维护任意产品；PROVIDER 仅本人 create_by（空/异主 → 403 ERR_FORBIDDEN；不存在 → 404）。
        无角色权限（USER）→ 403 ERR_MAINTENANCE_FORBIDDEN。
      operationId: updateMaintenanceEntry
      parameters:
        - $ref: "#/components/parameters/ProductId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MaintenanceAssociate"
      responses:
        "200":
          description: 已关联
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseMaintenanceEntry"
        "400":
          description: 非三级叶节点
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"
              examples:
                leafRequired:
                  value:
                    code: ERR_CATEGORY_LEAF_REQUIRED
                    message: 产品须挂载三级分类节点
                    data: null
                    correlationId: "corr-maint-leaf-1"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /catalog/maintenance/entries/batch:
    post:
      tags: [catalog]
      summary: 批量目录维护关联
      description: |
        ADMIN 可批量维护任意产品；PROVIDER 逐条校验 create_by=本人，越权条目进入 failures（ERR_FORBIDDEN）。
        无角色权限（USER）→ 403 ERR_MAINTENANCE_FORBIDDEN。
      operationId: batchUpdateMaintenanceEntries
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MaintenanceBatchAssociate"
      responses:
        "200":
          description: 批量结果
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseMaintenanceBatchResult"
        "403":
          $ref: "#/components/responses/Forbidden"
  /catalog/products/import/template:
    get:
      tags: [catalog]
      summary: 下载批量导入模板（xlsx/csv 列头）
      operationId: getImportTemplate
      parameters:
        - name: format
          in: query
          schema:
            type: string
            enum: [xlsx, csv]
            default: xlsx
      responses:
        "200":
          description: 模板文件
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
            text/csv:
              schema:
                type: string
                format: binary
  /catalog/products/import:
    post:
      tags: [catalog]
      summary: 同步批量导入产品（仅 PROVIDER；REQ-CAT-008；无异步 job）
      description: |
        导入仅 PROVIDER；ADMIN/USER → ERR_FORBIDDEN（403）。UI 宿主为「我的数据产品」。
        请求级失败（整单拒绝）：文件 >10MB → ERR_IMPORT_FILE_TOO_LARGE；
        无法解析/非 xlsx|csv → ERR_IMPORT_FORMAT_INVALID；
        可解析但列名行集合 ≠ v0729 权威 ImportTemplateColumns → ERR_IMPORT_TEMPLATE_UNSUPPORTED。
        HTTP 为错误响应，无部分写入；不得伪装为行级结果态。
        行级失败：文件已接受并处理完后返回成功 envelope + successCount/failureCount/reportId；
        failureCount>0 时 reportId 必填。ERR_IMPORT_ROW_INVALID 仅作报告内原因码。
        行业分类列语义：GB/T 4754 门类中文原文枚举（对齐 v0729 C 列 dataValidation，写入 industryCategory）；
        非法门类 → 行级失败（ERR_IMPORT_ROW_INVALID）。不再要求解析为三级叶子。
        模板无产品编码列时服务端按 DEC-WSC-001 自动生成。
      operationId: importProducts
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: ≤10MB；xlsx 或 csv
      responses:
        "200":
          description: 同步导入完成（可含部分行级失败）
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseImportResult"
              examples:
                allSuccess:
                  summary: 全成功
                  value:
                    code: "0"
                    message: OK
                    data:
                      successCount: 10
                      failureCount: 0
                      reportId: null
                    correlationId: "corr-imp-ok"
                partial:
                  summary: 部分成功（行级）
                  value:
                    code: "0"
                    message: OK
                    data:
                      successCount: 7
                      failureCount: 3
                      reportId: "rep-20260730-001"
                    correlationId: "corr-imp-partial"
                allRowFail:
                  summary: 全失败（行级）
                  value:
                    code: "0"
                    message: OK
                    data:
                      successCount: 0
                      failureCount: 5
                      reportId: "rep-20260730-002"
                    correlationId: "corr-imp-rows"
        "400":
          description: 请求级拒绝（整单无写入）
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"
              examples:
                tooLarge:
                  summary: 文件超限
                  value:
                    code: ERR_IMPORT_FILE_TOO_LARGE
                    message: 导入文件超过 10MB
                    data: null
                    correlationId: "corr-imp-size"
                badFormat:
                  summary: 格式无法解析
                  value:
                    code: ERR_IMPORT_FORMAT_INVALID
                    message: 无法解析文件或非 xlsx/csv
                    data: null
                    correlationId: "corr-imp-fmt"
                templateUnsupported:
                  summary: 旧模板/非 v0729 列名行（文件级拒绝）
                  value:
                    code: ERR_IMPORT_TEMPLATE_UNSUPPORTED
                    message: 导入模板不受支持：表头须为 v0729 权威列名行
                    data: null
                    correlationId: "corr-imp-tpl"
        "403":
          $ref: "#/components/responses/Forbidden"
  /catalog/products/import/reports/{reportId}:
    get:
      tags: [catalog]
      summary: 下载导入错误报告（TTL≤24h；白名单字段）
      operationId: getImportErrorReport
      parameters:
        - name: reportId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: 错误报告（行号+产品编码+原因码/文案）
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseImportErrorReport"
              examples:
                rows:
                  value:
                    code: "0"
                    message: OK
                    data:
                      reportId: "rep-20260730-001"
                      expiresAt: "2026-07-31T08:00:00Z"
                      rows:
                        - rowNumber: 4
                          productCode: "MED-EMR-0099"
                          reasonCode: ERR_IMPORT_ROW_INVALID
                          reasonMessage: 行业分类非法，须为 GB/T 4754 门类枚举：电子病历
                        - rowNumber: 8
                          productCode: "BAD-CODE"
                          reasonCode: ERR_PRODUCT_CODE_FORMAT
                          reasonMessage: 产品编码格式非法
                    correlationId: "corr-imp-rep"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: 过期或未知 reportId
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"
              examples:
                notFound:
                  value:
                    code: ERR_IMPORT_REPORT_NOT_FOUND
                    message: 错误报告不存在或已过期
                    data: null
                    correlationId: "corr-imp-rep-404"
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: WSC_SESSION
      description: |
        服务端会话 Cookie。登录后写入；登出立即失效。
        角色绑定在会话中；写 API 以服务端会话角色为准。
        角色/权限变更后须吊销或强制刷新旧会话。
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque-or-JWT
      description: |
        可选等价方案：Bearer 令牌必须映射可吊销的服务端会话表。
        非「纯无状态 JWT」；吊销后写 API 必须拒绝。
  parameters:
...
    ProductId:
      name: productId
      in: path
      required: true
      schema: { type: string }
  responses:
...
    Forbidden:
      description: 无权限
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiResponseError"
          examples:
            forbidden:
              value:
                code: ERR_FORBIDDEN
                message: 当前角色无权执行该写操作
                data: null
                correlationId: "corr-403"
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiResponseError"
  schemas:
    ApiEnvelope:
      type: object
      required: [code, message, data, correlationId]
      properties:
        code:
          type: string
          description: 成功为 "0"；业务错误为 ERR_* 或 HTTP 语义码字符串
        message:
          type: string
        data:
          nullable: true
        correlationId:
          type: string
...
    ApiResponseError:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          properties:
            code:
              type: string
              enum:
                - ERR_PRODUCT_CODE_FORMAT
                - ERR_PRODUCT_CODE_CONFLICT
                - ERR_CATEGORY_HAS_PRODUCTS
                - ERR_CATEGORY_LEAF_REQUIRED
                - ERR_MAINTENANCE_FORBIDDEN
                - ERR_IMPORT_FILE_TOO_LARGE
                - ERR_IMPORT_FORMAT_INVALID
                - ERR_IMPORT_TEMPLATE_UNSUPPORTED
                - ERR_IMPORT_REPORT_NOT_FOUND
                - ERR_FORBIDDEN
                - ERR_ROLE_SWITCH_DISABLED
                - ERR_USER_USERNAME_CONFLICT
                - ERR_USER_NOT_FOUND
                - "401"
                - "404"
            data:
              nullable: true
              description: ERR_CATEGORY_HAS_PRODUCTS 时含 reason、productCount
...
    ProductType:
      type: string
      enum: [DATASET, REPORT, API, OTHER]
      description: |
        DATASET=数据集, REPORT=数据报告, API=数据接口, OTHER=其他数据产品。
        OTHER：允许 typeSpecific.other（含 contentDescription 及共用时间/地域）；
        不得再要求 OTHER 必须为空对象。
...
    ProductCode:
      type: string
      pattern: '^[A-Z0-9]+-[A-Z0-9]+-[0-9]{4,}$'
      description: DEC-WSC-001 `{DOMAIN}-{FEATURE}-{NNNN}`
      example: MED-EMR-0002
    ApiEndpointParameter:
      type: object
      properties:
        name: { type: string }
        type: { type: string }
        required: { type: boolean }
        description: { type: string }
    ApiEndpointResponse:
      type: object
      properties:
        code: { type: string }
        description: { type: string }
    ApiEndpoint:
      type: object
      properties:
        id: { type: string }
        method: { type: string, description: "GET|POST|PUT|DELETE|PATCH 等" }
        path: { type: string }
        summary: { type: string }
        description: { type: string }
        parameters:
          type: array
          items: { $ref: "#/components/schemas/ApiEndpointParameter" }
        responses:
          type: array
          items: { $ref: "#/components/schemas/ApiEndpointResponse" }
        responseBodySchema: { type: string }
        requestBodySchema: { type: string }
    TypeSpecificApi:
      type: object
      description: |
        数据接口专属。归档真源 swaggerFileContent；派生存贮 endpoints[]。
        写冲突（二者均非空且不一致）时以客户端 endpoints 为准，禁止服务端静默重算覆盖。
        兼容旧字段 endpoint（非唯一真源）。
      properties:
        swaggerFileContent: { type: string, description: 原始 OpenAPI/Swagger 文本；可空 }
        endpoints:
          type: array
          items: { $ref: "#/components/schemas/ApiEndpoint" }
        fieldDescription: { type: string }
        dataSample: { type: string }
        timeRange: { type: string }
        regionScope: { type: string }
        endpoint: { type: string, description: 兼容旧字段；可读；写非唯一真源 }
    TypeSpecificDataset:
      type: object
      properties:
        timeRange: { type: string }
        regionScope: { type: string }
        dataScale: { type: string }
        dataForm: { type: string, description: 图片/文本/视频/音频/表格/其他 }
        fieldDescription: { type: string }
        dataSample: { type: string }
    TypeSpecificReportOrOther:
      type: object
      description: report 与 other 均允许 contentDescription + 共用时间/地域
      properties:
        timeRange: { type: string }
        regionScope: { type: string }
        contentDescription: { type: string }
    TypeSpecificFields:
      type: object
      description: |
        按 productType 择一填充 api / dataset / report / other。
        OTHER 允许 other.contentDescription 及共用时间/地域（修订 OQ-004）；
        不再要求 OTHER 必须为空对象。
      properties:
        dataset:
          $ref: "#/components/schemas/TypeSpecificDataset"
        report:
          $ref: "#/components/schemas/TypeSpecificReportOrOther"
        api:
          $ref: "#/components/schemas/TypeSpecificApi"
        other:
          $ref: "#/components/schemas/TypeSpecificReportOrOther"
...
    ChainVersion:
      type: object
      required: [versionId, versionNo, timestamp, metadataHash, ownerDID, certificate]
      properties:
        versionId: { type: string }
        versionNo: { type: integer }
        timestamp: { type: string, format: date-time }
        metadataHash:
          type: string
          description: 日志禁止明文或仅截断指纹
        ownerDID:
          type: string
          description: 日志禁止明文或仅截断指纹
        certificate:
          type: object
          required: [owner]
          properties:
            owner: { type: string }
            timestamp: { type: string, format: date-time }
    ApiResponseChainVersionList:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          properties:
            data:
              type: object
              properties:
                items:
                  type: array
                  items:
                    $ref: "#/components/schemas/ChainVersion"
    CatalogSnapshot:
      type: object
      description: |
        版本化不可变目录快照。字段覆盖 REQ-CAT-004 分组或文档化子集：
        基础信息、供应商、数据产权、类型专属（含 OTHER.contentDescription）、标签、简介、场景。
      required: [versionId, productCode, productName, productType, capturedAt]
      properties:
        versionId: { type: string }
        productCode: { $ref: "#/components/schemas/ProductCode" }
        productName: { type: string }
        productType: { $ref: "#/components/schemas/ProductType" }
        categoryPath:
          type: string
          description: 三级分类完整路径，如「可信空间/金融/征信」
        categoryPathParts:
          type: object
          description: 存证快照三级路径字段（L1/L2/L3）
          properties:
            l1: { type: string }
            l2: { type: string }
            l3: { type: string }
        basicInfo:
          type: object
          additionalProperties: true
        supplierInfo:
          type: object
          additionalProperties: true
        propertyRights:
          type: object
          additionalProperties: true
        typeSpecific:
          $ref: "#/components/schemas/TypeSpecificFields"
        tags:
          type: array
          items: { type: string }
        summary: { type: string }
        scenario: { type: string }
        capturedAt: { type: string, format: date-time }
        attestation:
          type: object
          properties:
            metadataHash: { type: string }
            ownerDID: { type: string }
            timestamp: { type: string, format: date-time }
            certificate:
              type: object
              properties:
                owner: { type: string }
    ApiResponseChainSnapshot:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/CatalogSnapshot"
    MaintenanceStatus:
      type: string
      enum: [PENDING, MAINTAINED]
      description: 待关联 / 已维护；目录维护条目 ≡ 数据产品
    MaintenanceEntry:
      type: object
      description: 与 Product 同一资源 id/productCode（不得拆两套冲突实体）
      required: [id, productCode, productName, maintenanceStatus]
      properties:
        id: { type: string }
        productCode: { $ref: "#/components/schemas/ProductCode" }
        productName: { type: string }
        maintenanceStatus: { $ref: "#/components/schemas/MaintenanceStatus" }
        l3CategoryId: { type: string, nullable: true }
        categoryPath: { type: string, nullable: true }
    MaintenancePage:
      type: object
      required: [items, page, pageSize, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/MaintenanceEntry"
        page: { type: integer }
        pageSize: { type: integer }
        total: { type: integer, format: int64 }
    MaintenanceAssociate:
      type: object
      required: [l3CategoryId]
      properties:
        l3CategoryId:
          type: string
          description: 目标三级节点 id
    MaintenanceBatchAssociate:
      type: object
      required: [productIds, l3CategoryId]
      properties:
        productIds:
          type: array
          items: { type: string }
          minItems: 1
        l3CategoryId: { type: string }
    MaintenanceBatchResult:
      type: object
      required: [successCount, failureCount]
      properties:
        successCount: { type: integer }
        failureCount: { type: integer }
        failures:
          type: array
          items:
            type: object
            properties:
              productId: { type: string }
              reasonCode: { type: string }
              reasonMessage: { type: string }
    ApiResponseMaintenanceEntry:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/MaintenanceEntry"
    ApiResponseMaintenancePage:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/MaintenancePage"
    ApiResponseMaintenanceBatchResult:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/MaintenanceBatchResult"
    ImportResult:
      type: object
      required: [successCount, failureCount]
      properties:
        successCount: { type: integer }
        failureCount: { type: integer }
        reportId:
          type: string
          nullable: true
          description: failureCount>0 时必填；全成功可空
...
    ImportErrorRow:
      type: object
      required: [rowNumber, reasonCode, reasonMessage]
      properties:
        rowNumber: { type: integer }
        productCode: { type: string, nullable: true }
        reasonCode:
          type: string
          description: 报告内原因码；含 ERR_IMPORT_ROW_INVALID（不作整单 HTTP code）
        reasonMessage: { type: string }
    ImportErrorReport:
      type: object
      required: [reportId, expiresAt, rows]
      properties:
        reportId: { type: string }
        expiresAt: { type: string, format: date-time }
        rows:
          type: array
          items:
            $ref: "#/components/schemas/ImportErrorRow"
    ApiResponseImportResult:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/ImportResult"
    ApiResponseImportErrorReport:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/ImportErrorReport"